finish relationship product and sell and routes

+++ app/Http/Requests/api/v1/SellUpdateRequest.php
@@ -24,7 +24,7 @@ class SellUpdateRequest extends FormRequest
         return [
             //
             'user_id' => 'exists:users,id',
-            'total_price' => 'float',
+            'total_price' => 'numeric',
             'status' => 'string',
         ];
     }
